refactor: SkillsSection File Structure
