+++ src/components/SkillsSection/index.tsx
@@ -1,0 +1,3 @@
+export { Skills } from './skills';
+export { EchartRosePieChart } from './echartRosePieChart';
+export { EchartsLoader } from './echartsLoader';
